requirements.txt
openai
slack-bolt
sympy
python-dotenv
